+++ test/general-syntax-rei.t/run.t
@@ -73,3 +73,18 @@ Format general interface syntax
         )
     ) =>
     reactClass;
+  
+  open%foo Bar;
+  
+  open! %foo Bar;
+  
+  type%foo t = int;
+  
+  type%x foo +=
+    | Int;
+  
+  module type x = {
+    let a: 'a. 'a => unit;
+  };
+  
+  let a: 'a. 'a => unit;