+++ test/mlFunctions.t/input.ml
@@ -1,0 +1,7 @@
+(* Copyright (c) 2015-present, Facebook, Inc. All rights reserved. *)
+
+let x =
+  ignore (fun y ->
+    let y = 4 in
+    y)
+